added some unit tests
